align exec scripts install service script + make it work with virtualenvs cmd do not need to be part of config remove piTFT exec script (obsolete + in documentation already)
