+++ modules/utils/cmd.py
@@ -1,0 +1,27 @@
+import subprocess
+
+from logger import app_logger
+
+
+def exec_cmd(cmd, cmd_print=True):
+    if cmd_print:
+        app_logger.info(cmd)
+    try:
+        subprocess.run(cmd)
+    except Exception:  # noqa
+        app_logger.exception(f"Failed executing {cmd}")
+
+
+def exec_cmd_return_value(cmd, cmd_print=True):
+    if cmd_print:
+        app_logger.info(cmd)
+    try:
+        p = subprocess.run(
+            cmd,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+        )
+        string = p.stdout.decode("utf8").strip()
+        return string
+    except Exception:  # noqa
+        app_logger.exception(f"Failed executing {cmd}")
