+++ scripts/install/etc/default/gpsd
