+++ scripts/install/etc/systemd/system/pizero_bikecomputer.service
@@ -6,14 +6,14 @@ Description=PizeroBikecomputer
 #After=gpsd.service local-fs.target
 
 [Service]
-WorkingDirectory=/home/pi/pizero_bikecomputer
-ExecStart=/home/pi/pizero_bikecomputer/exec-service.sh
+WorkingDirectory=
+ExecStart=
 
 Restart=no
 Type=simple
 
-User=pi
-Group=pi
+User=
+Group=
 
 [Install]
 WantedBy=multi-user.target
