+++ scripts/install/etc/systemd/system/pizero_bikecomputer_shutdown.service
