+++ scripts/install/usr/local/bin/disable-pitft
