+++ scripts/install/usr/local/bin/enable-pitft
