+++ scripts/install/usr/local/bin/pizero_bikecomputer_shutdown
